- now all mapping is done in Limiri so plugin is only for broadcasting over dbus - use nativeScanCode if Mir/QtMir does not map the input event to a known key
